+++ packages/ispect/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 4.0.5
+- Display of errors has been changed. Now, if the errors are large, only the first 10,000 characters will be displayed.
+This is done in order not to overload the widget and not to cause the application to hang.
+
 ## 4.0.4
 - Fixed the log description filtering method.